acm_pokerbot/requirements.txt
dotenv==0.9.9
websocket-client==1.8.0